chore: cleanup and consolidation
