--- packages/eslint-plugin-angular-template/src/configs/base.config.js
@@ -1,10 +1,0 @@
-export default [
-  {
-    languageOptions: {
-      parser: (await import('@angular-eslint/template-parser')).default,
-    },
-    plugins: {
-      '@angular-eslint/template': (await import('@angular-eslint/eslint-plugin-template')).default,
-    }
-  }
-];
