+++ packages/eslint-plugin-angular-template/src/configs/recommended.config.js
@@ -1,13 +1,19 @@
-import base from './base.config.js';
+import templatePlugin from '@angular-eslint/eslint-plugin-template';
 
 export default [
-  ...base,
   {
-    rules: {}
-  },
-  {
-    extends: [
-      'plugin:@angular-eslint/template/recommended'
-    ]
+    languageOptions: {
+      parser: templateParser,
+      parserOptions: {
+        tsconfigRootDir: __dirname
+      }
+    },
+    plugins: {
+      '@angular-eslint/template': templatePlugin,
+    },
+    rules: {
+      // Include all the recommended rules from @angular-eslint/template
+      ...templatePlugin.configs.recommended.rules
+    }
   }
 ];
