+++ packages/eslint-plugin/package.json
@@ -2,10 +2,15 @@
   "name": "@tylertech-eslint/eslint-plugin",
   "version": "1.0.0",
   "description": "Common TypeScript lint rules/preferences for Tyler Technologies. Compatible with ESLint v9+ (Flat Config only).",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js",
+    "./configs/recommended": "./dist/configs/recommended.js"
+  },
   "main": "dist/index.js",
   "typings": "typings/index.d.ts",
   "scripts": {
-    "build": "rimraf ./dist && microbundle --tsconfig tsconfig.build.json --no-sourcemap --target=node --compress --format=cjs && ncp ./src/configs ./dist/configs",
+    "build": "rimraf ./dist && microbundle --tsconfig tsconfig.build.json --no-sourcemap --target=node --compress --format=esm && ncp ./src/configs ./dist/configs",
     "build:publish": "rimraf ./publish && npm run build && gulp publish",
     "test": "jest --coverage",
     "gulp": "gulp",
