+++ packages/eslint-plugin/src/index.ts
@@ -1,15 +1,8 @@
+// ESLint v9 compatible index file
+import { rules } from './rules/index';
+
 // @ts-ignore - Import JS file in TypeScript
-import recommended from './configs/recommended.config.js';
+import recommended from './configs/recommended.js';
 
-import invalidRelativeImportPrefix, { RULE_NAME as invalidRelativeImportPrefixRuleName } from './rules/invalid-relative-Import-prefix';
-import requirePrivateUnderscore, { RULE_NAME as requirePrivateUnderscoreRuleName } from './rules/require-private-underscore';
-import requirePrivateModifier, { RULE_NAME as requirePrivateModifierRuleName } from './rules/require-private-modifier';
-
-
-export const rules = {
-    [requirePrivateModifierRuleName]: requirePrivateModifier,
-    [requirePrivateUnderscoreRuleName]: requirePrivateUnderscore,
-    [invalidRelativeImportPrefixRuleName]: invalidRelativeImportPrefix
-};
-
-export { recommended };
+// Export all components for ESLint v9 compatibility
+export { rules, recommended };
